Move hash into its own module
