+++ src/historian.rs
@@ -5,7 +5,8 @@ use std::thread::{spawn, JoinHandle};
 use std::collections::HashSet;
 use std::sync::mpsc::{sync_channel, Receiver, SyncSender};
 use std::time::Instant;
-use log::{hash, Entry, Sha256Hash};
+use hash::{hash, Sha256Hash};
+use log::Entry;
 use logger::{ExitReason, Logger};
 use signature::Signature;
 use event::Event;
