+++ src/log.rs
@@ -13,14 +13,10 @@
 /// fastest processor. Duration should therefore be estimated by assuming that the hash
 /// was generated by the fastest processor at the time the entry was logged.
 
-use generic_array::GenericArray;
-use generic_array::typenum::U32;
+use hash::{extend_and_hash, hash, Sha256Hash};
 use serde::Serialize;
 use event::Event;
-use sha2::{Digest, Sha256};
 use rayon::prelude::*;
-
-pub type Sha256Hash = GenericArray<u8, U32>;
 
 #[derive(Serialize, Deserialize, Debug, PartialEq, Eq, Clone)]
 pub struct Entry<T> {
@@ -48,20 +44,6 @@ impl<T: Serialize> Entry<T> {
         }
         self.id == next_hash(start_hash, self.num_hashes, &self.event)
     }
-}
-
-/// Return a Sha256 hash for the given data.
-pub fn hash(val: &[u8]) -> Sha256Hash {
-    let mut hasher = Sha256::default();
-    hasher.input(val);
-    hasher.result()
-}
-
-/// Return the hash of the given hash extended with the given value.
-pub fn extend_and_hash(id: &Sha256Hash, val: &[u8]) -> Sha256Hash {
-    let mut hash_data = id.to_vec();
-    hash_data.extend_from_slice(val);
-    hash(&hash_data)
 }
 
 /// Creates the hash 'num_hashes' after start_hash. If the event contains
